Split API part from command line part
